+++ cluster_artefacts/prometheus.tf
@@ -5,7 +5,7 @@ resource "helm_release" "prometheus" {
   version    = "8.15.1"
 
   values = [
-    templatefile("${path.module}/prometheus.yaml.tpl", { hostname = format("prometheus.%s.%s", var.cluster_name, var.tld) })
+    templatefile("${path.module}/prometheus.yaml.tpl", { hostname = format("prometheus.%s", var.tld) })
   ]
   count = var.prometheus ? 1 : 0
 }