+++ main.tf
@@ -1,6 +1,5 @@
 module "kubernetes_resources" {
   source            = "./cluster_artefacts"
-  cluster_name      = var.cluster_name
   email             = var.email
   tld               = var.tld
   velero            = var.velero