+++ variables.tf
@@ -1,8 +1,3 @@
-variable "cluster_name" {
-  description = "Name of the Kubernetes cluster."
-  type        = string
-}
-
 variable "cluster_uuid" {
   type = string
 }